feat: colours constants and makes small refactor of constants and utils
